finalproject/requirements.txt
argparse
torchvision
pillow
numpy